requirements.txt
streamlit==1.28.0
datasets==2.14.0
transformers==4.34.0
sentence-transformers==2.2.2
torch==2.0.1
faiss-cpu==1.7.4
pandas==2.0.3
numpy==1.24.3
openai==0.28.0
ragas==0.0.18
langchain==0.0.292
chromadb==0.4.15
plotly==5.17.0
scikit-learn==1.3.0